# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/v1/user": {
            "get": {
                "description": "get  all users by where",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "get all users by",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.CommonResponse-array_user_ResponseUser"
                        }
                    }
                }
            },
            "put": {
                "description": "update userinfo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userinfo"
                ],
                "summary": "update userinfo",
                "parameters": [
                    {
                        "description": "JSON Data",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.CommonResponseBuilder-user_ResponseUser"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create suer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user create"
                ],
                "summary": "create user",
                "parameters": [
                    {
                        "description": "JSON Data",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.NewUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CommonResponseBuilder"
                        }
                    }
                }
            }
        },
        "/v1/user/search": {
            "get": {
                "description": "search users by query",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search users"
                ],
                "summary": "search users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "PageSize",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sortBy",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sortDirection",
                        "name": "sortDirection",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "status",
                        "name": "status_match",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "userName",
                        "name": "user_name_like",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PageList-array_user_ResponseUser"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CommonResponseBuilder": {
            "type": "object"
        },
        "controllers.CommonResponseBuilder-user_ResponseUser": {
            "type": "object"
        },
        "domain.CommonResponse-array_user_ResponseUser": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.ResponseUser"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "domain.CustomTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
        "domain.DataFilters": {
            "type": "object",
            "properties": {
                "dataRanges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.DateRangeFilter"
                    }
                },
                "likeFilters": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "matches": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "sortBy": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sortDirection": {
                    "$ref": "#/definitions/domain.SortDirection"
                }
            }
        },
        "domain.DateRangeFilter": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            }
        },
        "domain.PageList-array_user_ResponseUser": {
            "type": "object",
            "properties": {
                "filters": {
                    "$ref": "#/definitions/domain.DataFilters"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.ResponseUser"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "domain.SortDirection": {
            "type": "string",
            "enum": [
                "asc",
                "desc"
            ],
            "x-enum-varnames": [
                "SortAsc",
                "SortDesc"
            ]
        },
        "user.NewUserRequest": {
            "type": "object",
            "required": [
                "email",
                "user_name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "header_img": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nick_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "user.ResponseUser": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/domain.CustomTime"
                },
                "email": {
                    "type": "string"
                },
                "header_img": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nick_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "updated_at": {
                    "$ref": "#/definitions/domain.CustomTime"
                },
                "user_name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}